requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli==1.27.38
flake8
python-dotenv>=0.5.1
pytorch
matplotlib
numpy
tqdm
scipy